refactor: Improve the encoder api
aacEncCode has to be called repeatedly to encode a buffer of input, and
this pushes a lot of book keeping overahead onto the caller. This
changes the method to accept a std::io::Read and std::io::Write instead,
which allow the caller to use a std::io::Cursor or any other convenient
implementation of those standard traits.